server/requirements.txt
anyio==3.6.1
Arpeggio==1.10.2
asgiref==3.5.2
attrs==21.4.0
boto3==1.24.30
botocore==1.27.30
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
dill==0.3.5.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
fastapi-cloudauth==0.4.3
faunadb==4.2.0
future==0.18.2
greenlet==1.1.2
grpcio==1.47.0
h11==0.13.0
h2==2.6.2
hpack==3.0.0
httptools==0.4.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.3
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
orjson==3.7.7
parver==0.3.1
protobuf==4.21.2
psycopg2-binary==2.9.3
pulumi==3.36.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
qrcode==7.3.1
requests==2.28.1
rsa==4.8
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
starlette==0.19.1
typing-extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
